+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 - Extension settings reorganized for clarity, easiness of setup, and additional customizations
 - Performance improvements
 - Improved instruction support within macro definitions. #134
-- TBD Better macro matchers. #134
 - New `z80-asm-meter.timing.executionFlow.stopOnUnconditionalJump` and `z80-asm-meter.timing.atExit.stopOnUnconditionalJump` settings. #143
 - Timing hints can now ignore empty lines with commented out source code. #144
+- RegExp-based user-defined timing hints. #134
 
 ## 4.3.0